Windows compat. mods
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@551 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ c++/nitf/include/nitf/RESubheader.hpp
@@ -88,7 +88,7 @@ public:
     char * getSubheaderFields() const;
 
     //! Get the dataLength
-    nitf::Uint32 getDataLength() const;
+    nitf::Uint64 getDataLength() const;
 
     //! Set the dataLength
     void setDataLength(nitf::Uint32 value);
